requirements.txt
blinker==1.6.2
click==8.0.1
colorama==0.4.6
DateTime==5.2
dnspython==1.16.0
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-WTF==0.15.1
gunicorn==21.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pymongo==3.12.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.0.1
WTForms==2.3.3
zope.interface==6.0
